requirements.txt
python-dotenv
datasets==3.5.0
aiohttp==3.11.15
openai==1.70.0
instructor==1.7.9
diskcache==5.6.3
tabulate==0.9.0
tiktoken    
pip install pydantic-ai
urllib2